Fix Exchange

+++ Models/VoucherDetail.cs
@@ -11,10 +11,10 @@ namespace AccountingSystem.Models
         public int numberVoucher { get; set; }
         public string gloss { get; set; }
         public string account{ get; set; }
-        public decimal amountOwed{ get; set; }
-        public decimal amountAssets{ get; set; }
-        public decimal amountOwedAlt{ get; set; }
-        public decimal amountAssetsAlt{ get; set; }
+        public double amountOwed{ get; set; }
+        public double amountAssets{ get; set; }
+        public double  amountOwedAlt{ get; set; }
+        public double amountAssetsAlt{ get; set; }
         public int idUser{ get; set; }
         public int stateAddorUpdate { get; set; } = 1;
         public int  idVoucher{ get; set; }
